Let's begin by separating all the formatting functions out
