+++ lib/Client.js
@@ -28,6 +28,9 @@ const User = require("./structures/User");
 const VoiceConnectionManager = require("./voice/VoiceConnectionManager");
 const AutoModerationRule = require("./structures/AutoModerationRule");
 const emitDeprecation = require("./util/emitDeprecation");
+const formatAllowedMentions = require("./util/formatAllowedMentions");
+const processAttachments = require("./util/processAttachments");
+
 
 let EventEmitter;
 try {
@@ -170,21 +173,8 @@ class Client extends EventEmitter {
          * Dysnomia options
          * @type {Object}
          */
-        this.options = Object.assign({
-            allowedMentions: {
-                users: true,
-                roles: true
-            },
-            defaultImageFormat: "jpg",
-            defaultImageSize: 128,
-            messageLimit: 100,
-            opusOnly: false,
-            rest: {},
-            restMode: false,
-            ws: {},
-            gateway: {}
-        }, options);
-        this.options.allowedMentions = this._formatAllowedMentions(this.options.allowedMentions);
+        this.options = Object.assign({}, Constants.DefaultClientOptions, options);
+        this.options.allowedMentions = formatAllowedMentions(this.options.allowedMentions, this);
         if(!Constants.ImageFormats.includes(this.options.defaultImageFormat.toLowerCase())) {
             throw new TypeError(`Invalid default image format: ${this.options.defaultImageFormat}`);
         }
@@ -797,7 +787,7 @@ class Client extends EventEmitter {
             } else if(content.content !== undefined && typeof content.content !== "string") {
                 content.content = "" + content.content;
             }
-            content.allowed_mentions = this._formatAllowedMentions(content.allowedMentions);
+            content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this);
             content.sticker_ids = content.stickerIDs;
             if(content.messageReference) {
                 content.message_reference = content.messageReference;
@@ -820,7 +810,7 @@ class Client extends EventEmitter {
             }
         }
 
-        const {files, attachments} = this._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.requestHandler.request("POST", Endpoints.CHANNEL_MESSAGES(channelID), true, content, files).then((message) => new Message(message, this));
@@ -913,11 +903,11 @@ class Client extends EventEmitter {
             if(options.message.content !== undefined && typeof options.message.content !== "string") {
                 options.message.content = "" + options.message.content;
             }
-            options.message.allowed_mentions = this._formatAllowedMentions(options.message.allowedMentions);
+            options.message.allowed_mentions = formatAllowedMentions(options.message.allowedMentions, this);
             options.message.sticker_ids = options.message.stickerIDs;
         }
 
-        const {files, attachments} = options.message ? this._processAttachments(options.message.attachments) : {};
+        const {files, attachments} = options.message ? processAttachments(options.message.attachments) : {};
         if(options.message) {
             options.message.attachments = attachments;
         }
@@ -1734,11 +1724,11 @@ class Client extends EventEmitter {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this);
             }
         }
 
-        const {files, attachments} = content.attachments ? this._processAttachments(content.attachments) : [];
+        const {files, attachments} = content.attachments ? processAttachments(content.attachments) : {};
         content.attachments = attachments;
 
         return this.requestHandler.request("PATCH", Endpoints.CHANNEL_MESSAGE(channelID, messageID), true, content, files).then((message) => new Message(message, this));
@@ -1920,10 +1910,10 @@ class Client extends EventEmitter {
             qs += "&thread_id=" + options.threadID;
         }
         if(options.allowedMentions) {
-            options.allowed_mentions = this._formatAllowedMentions(options.allowedMentions);
+            options.allowed_mentions = formatAllowedMentions(options.allowedMentions, this);
         }
 
-        const {files, attachments} = options.attachments ? this._processAttachments(options.attachments) : [];
+        const {files, attachments} = options.attachments ? processAttachments(options.attachments) : {};
         options.attachments = attachments;
 
         return this.requestHandler.request("PATCH", Endpoints.WEBHOOK_MESSAGE(webhookID, token, messageID) + (qs ? "?" + qs : ""), false, options, files).then((response) => new Message(response, this));
@@ -1991,7 +1981,7 @@ class Client extends EventEmitter {
             qs += "&thread_id=" + options.threadID;
         }
 
-        const {files, attachments} = options.attachments ? this._processAttachments(options.attachments) : [];
+        const {files, attachments} = options.attachments ? processAttachments(options.attachments) : {};
 
         return this.requestHandler.request("POST", Endpoints.WEBHOOK_TOKEN(webhookID, token) + (qs ? "?" + qs : ""), !!options.auth, {
             content: options.content,
@@ -2000,7 +1990,7 @@ class Client extends EventEmitter {
             avatar_url: options.avatarURL,
             tts: options.tts,
             flags: options.flags,
-            allowed_mentions: this._formatAllowedMentions(options.allowedMentions),
+            allowed_mentions: formatAllowedMentions(options.allowedMentions, this),
             components: options.components,
             attachments: attachments,
             thread_name: options.threadName
@@ -3179,79 +3169,6 @@ class Client extends EventEmitter {
     */
     unpinMessage(channelID, messageID) {
         return this.requestHandler.request("DELETE", Endpoints.CHANNEL_PIN(channelID, messageID), true);
-    }
-
-    _formatAllowedMentions(allowed) {
-        if(!allowed) {
-            return this.options.allowedMentions;
-        }
-        const result = {
-            parse: []
-        };
-        if(allowed.everyone) {
-            result.parse.push("everyone");
-        }
-        if(allowed.roles === true) {
-            result.parse.push("roles");
-        } else if(Array.isArray(allowed.roles)) {
-            if(allowed.roles.length > 100) {
-                throw new Error("Allowed role mentions cannot exceed 100.");
-            }
-            result.roles = allowed.roles;
-        }
-        if(allowed.users === true) {
-            result.parse.push("users");
-        } else if(Array.isArray(allowed.users)) {
-            if(allowed.users.length > 100) {
-                throw new Error("Allowed user mentions cannot exceed 100.");
-            }
-            result.users = allowed.users;
-        }
-        if(allowed.repliedUser !== undefined) {
-            result.replied_user = allowed.repliedUser;
-        }
-        return result;
-    }
-
-    _formatImage(url, format, size) {
-        if(!format || !Constants.ImageFormats.includes(format.toLowerCase())) {
-            format = url.includes("/a_") ? "gif" : this.options.defaultImageFormat;
-        }
-        if(!size || size < Constants.ImageSizeBoundaries.MINIMUM || size > Constants.ImageSizeBoundaries.MAXIMUM || (size & (size - 1))) {
-            size = this.options.defaultImageSize;
-        }
-        return `${Endpoints.CDN_URL}${url}.${format}?size=${size}`;
-    }
-
-    _processAttachments(attachments) {
-        if(!attachments) {
-            return {};
-        }
-        const files = [];
-        const resultAttachments = [];
-
-        attachments.forEach((attachment, idx) => {
-            if(attachment.id) {
-                resultAttachments.push(attachment);
-            } else {
-                files.push({
-                    fieldName: `files[${idx}]`,
-                    file: attachment.file,
-                    name: attachment.filename
-                });
-
-                resultAttachments.push({
-                    ...attachment,
-                    file: undefined,
-                    id: idx
-                });
-            }
-        });
-
-        return {
-            files: files.length ? files : undefined,
-            attachments: resultAttachments
-        };
     }
 
     toString() {
