+++ lib/Constants.js
@@ -218,6 +218,21 @@ module.exports.ConnectionVisibilityTypes = {
     EVERYONE: 1
 };
 
+module.exports.DefaultClientOptions = Object.freeze({
+    allowedMentions: {
+        users: true,
+        roles: true
+    },
+    defaultImageFormat: "jpg",
+    defaultImageSize: 128,
+    messageLimit: 100,
+    opusOnly: false,
+    rest: {},
+    restMode: false,
+    ws: {},
+    gateway: {}
+});
+
 module.exports.DefaultMessageNotificationLevels = {
     ALL_MESSAGES:  0,
     ONLY_MENTIONS: 1
