+++ lib/structures/CommandInteraction.js
@@ -8,8 +8,10 @@ const Channel = require("./Channel");
 const Message = require("./Message");
 const Attachment = require("./Attachment");
 const Collection = require("../util/Collection");
+const formatAllowedMentions = require("../util/formatAllowedMentions");
 
 const {InteractionResponseTypes} = require("../Constants");
+const processAttachments = require("../util/processAttachments");
 
 /**
 * Represents an application command interaction.
@@ -164,11 +166,11 @@ class CommandInteraction extends Interaction {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this.#client._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this.#client);
             }
         }
 
-        const {files, attachments} = this.#client._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.#client.createInteractionResponse.call(this.#client, this.id, this.token, {
