+++ lib/structures/ComponentInteraction.js
@@ -7,8 +7,10 @@ const Role = require("./Role");
 const Channel = require("./Channel");
 const Message = require("./Message");
 const Collection = require("../util/Collection");
+const formatAllowedMentions = require("../util/formatAllowedMentions");
 
 const {InteractionResponseTypes} = require("../Constants");
+const processAttachments = require("../util/processAttachments");
 
 /**
 * Represents a message component interaction.
@@ -153,11 +155,11 @@ class ComponentInteraction extends Interaction {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this.#client._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this.#client);
             }
         }
 
-        const {files, attachments} = this.#client._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.#client.createInteractionResponse.call(this.#client, this.id, this.token, {
@@ -341,11 +343,11 @@ class ComponentInteraction extends Interaction {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this.#client._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this.#client);
             }
         }
 
-        const {files, attachments} = this.#client._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.#client.createInteractionResponse.call(this.#client, this.id, this.token, {
