+++ lib/structures/Guild.js
@@ -13,6 +13,7 @@ const GuildScheduledEvent = require("./GuildScheduledEvent");
 const {Permissions} = require("../Constants");
 const StageInstance = require("./StageInstance");
 const ThreadChannel = require("./ThreadChannel");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a guild
@@ -457,7 +458,7 @@ class Guild extends Base {
      * @type {String?}
      */
     get bannerURL() {
-        return this.banner ? this.#client._formatImage(Endpoints.BANNER(this.id, this.banner)) : null;
+        return this.banner ? formatImage(Endpoints.BANNER(this.id, this.banner), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -465,7 +466,7 @@ class Guild extends Base {
      * @type {String?}
      */
     get iconURL() {
-        return this.icon ? this.#client._formatImage(Endpoints.GUILD_ICON(this.id, this.icon)) : null;
+        return this.icon ? formatImage(Endpoints.GUILD_ICON(this.id, this.icon), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -473,7 +474,7 @@ class Guild extends Base {
      * @type {String?}
      */
     get splashURL() {
-        return this.splash ? this.#client._formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash)) : null;
+        return this.splash ? formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -481,7 +482,7 @@ class Guild extends Base {
      * @type {String?}
      */
     get discoverySplashURL() {
-        return this.discoverySplash ? this.#client._formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash)) : null;
+        return this.discoverySplash ? formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -768,7 +769,7 @@ class Guild extends Base {
     * @returns {String?}
     */
     dynamicBannerURL(format, size) {
-        return this.banner ? this.#client._formatImage(Endpoints.BANNER(this.id, this.banner), format, size) : null;
+        return this.banner ? formatImage(Endpoints.BANNER(this.id, this.banner), format, size, this.#client) : null;
     }
 
     /**
@@ -778,7 +779,7 @@ class Guild extends Base {
      * @returns {String?}
      */
     dynamicDiscoverySplashURL(format, size) {
-        return this.discoverySplash ? this.#client._formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), format, size) : null;
+        return this.discoverySplash ? formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), format, size, this.#client) : null;
     }
 
     /**
@@ -788,7 +789,7 @@ class Guild extends Base {
     * @returns {String?}
     */
     dynamicIconURL(format, size) {
-        return this.icon ? this.#client._formatImage(Endpoints.GUILD_ICON(this.id, this.icon), format, size) : null;
+        return this.icon ? formatImage(Endpoints.GUILD_ICON(this.id, this.icon), format, size, this.#client) : null;
     }
 
     /**
@@ -798,7 +799,7 @@ class Guild extends Base {
     * @returns {String?}
     */
     dynamicSplashURL(format, size) {
-        return this.splash ? this.#client._formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), format, size) : null;
+        return this.splash ? formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), format, size, this.#client) : null;
     }
 
     /**
