+++ lib/structures/GuildPreview.js
@@ -2,6 +2,7 @@
 
 const Base = require("./Base");
 const Endpoints = require("../rest/Endpoints.js");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a GuildPreview structure
@@ -74,7 +75,7 @@ class GuildPreview extends Base {
      * @type {String?}
      */
     get iconURL() {
-        return this.icon ? this.#client._formatImage(Endpoints.GUILD_ICON(this.id, this.icon)) : null;
+        return this.icon ? formatImage(Endpoints.GUILD_ICON(this.id, this.icon), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -82,7 +83,7 @@ class GuildPreview extends Base {
      * @type {String?}
      */
     get splashURL() {
-        return this.splash ? this.#client._formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash)) : null;
+        return this.splash ? formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -90,7 +91,7 @@ class GuildPreview extends Base {
      * @type {String?}
      */
     get discoverySplashURL() {
-        return this.discoverySplash ? this.#client._formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash)) : null;
+        return this.discoverySplash ? formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), undefined, undefined, this.#client) : null;
     }
 
     /**
@@ -100,7 +101,7 @@ class GuildPreview extends Base {
     * @returns {String?}
     */
     dynamicDiscoverySplashURL(format, size) {
-        return this.discoverySplash ? this.#client._formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), format, size) : null;
+        return this.discoverySplash ? formatImage(Endpoints.GUILD_DISCOVERY_SPLASH(this.id, this.discoverySplash), format, size, this.#client) : null;
     }
 
     /**
@@ -110,7 +111,7 @@ class GuildPreview extends Base {
     * @returns {String?}
     */
     dynamicIconURL(format, size) {
-        return this.icon ? this.#client._formatImage(Endpoints.GUILD_ICON(this.id, this.icon), format, size) : null;
+        return this.icon ? formatImage(Endpoints.GUILD_ICON(this.id, this.icon), format, size, this.#client) : null;
     }
 
     /**
@@ -120,7 +121,7 @@ class GuildPreview extends Base {
     * @returns {String?}
     */
     dynamicSplashURL(format, size) {
-        return this.splash ? this.#client._formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), format, size) : null;
+        return this.splash ? formatImage(Endpoints.GUILD_SPLASH(this.id, this.splash), format, size, this.#client) : null;
     }
 
     toJSON(props = []) {
