+++ lib/structures/GuildScheduledEvent.js
@@ -2,6 +2,7 @@
 
 const Base = require("./Base");
 const Endpoints = require("../rest/Endpoints");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a guild scheduled event
@@ -133,7 +134,7 @@ class GuildScheduledEvent extends Base {
      * @type {String?}
      */
     get imageURL() {
-        return this.image ? this.#client._formatImage(Endpoints.GUILD_SCHEDULED_EVENT_COVER(this.id, this.image)) : null;
+        return this.image ? formatImage(Endpoints.GUILD_SCHEDULED_EVENT_COVER(this.id, this.image), undefined, undefined, this.#client) : null;
     }
 
     /**
