+++ lib/structures/Member.js
@@ -4,6 +4,7 @@ const Base = require("./Base");
 const Endpoints = require("../rest/Endpoints");
 const User = require("./User");
 const VoiceState = require("./VoiceState");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a server member
@@ -158,7 +159,7 @@ class Member extends Base {
      * @type {String}
      */
     get avatarURL() {
-        return this.avatar ? this.guild.shard.client._formatImage(Endpoints.GUILD_AVATAR(this.guild.id, this.id, this.avatar)) : this.user.avatarURL;
+        return this.avatar ? formatImage(Endpoints.GUILD_AVATAR(this.guild.id, this.id, this.avatar), undefined, undefined, this.guild.shard.client) : this.user.avatarURL;
     }
 
     /**
@@ -305,7 +306,7 @@ class Member extends Base {
         if(!this.avatar) {
             return this.user.dynamicAvatarURL(format, size);
         }
-        return this.guild.shard.client._formatImage(Endpoints.GUILD_AVATAR(this.guild.id, this.id, this.avatar), format, size);
+        return formatImage(Endpoints.GUILD_AVATAR(this.guild.id, this.id, this.avatar), format, size, this.guild.shard.client);
     }
     /**
     * Edit the guild member
