+++ lib/structures/ModalSubmitInteraction.js
@@ -1,8 +1,10 @@
 "use strict";
 
 const Interaction = require("./Interaction");
+const formatAllowedMentions = require("../util/formatAllowedMentions");
 
 const {InteractionResponseTypes} = require("../Constants");
+const processAttachments = require("../util/processAttachments");
 
 /**
 * Represents a modal submit interaction.
@@ -98,11 +100,11 @@ class ModalSubmitInteraction extends Interaction {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this.#client._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this.#client);
             }
         }
 
-        const {files, attachments} = this.#client._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.#client.createInteractionResponse.call(this.#client, this.id, this.token, {
@@ -270,11 +272,11 @@ class ModalSubmitInteraction extends Interaction {
                 content.content = "" + content.content;
             }
             if(content.content !== undefined || content.embeds || content.allowedMentions) {
-                content.allowed_mentions = this.#client._formatAllowedMentions(content.allowedMentions);
+                content.allowed_mentions = formatAllowedMentions(content.allowedMentions, this.#client);
             }
         }
 
-        const {files, attachments} = this.#client._processAttachments(content.attachments);
+        const {files, attachments} = processAttachments(content.attachments);
         content.attachments = attachments;
 
         return this.#client.createInteractionResponse.call(this.#client, this.id, this.token, {
