+++ lib/structures/Role.js
@@ -3,6 +3,7 @@
 const Base = require("./Base");
 const Endpoints = require("../rest/Endpoints");
 const Permission = require("./Permission");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a role
@@ -124,7 +125,7 @@ class Role extends Base {
      * @type {String}
      */
     get iconURL() {
-        return this.icon ? this.guild.shard.client._formatImage(Endpoints.ROLE_ICON(this.id, this.icon)) : null;
+        return this.icon ? formatImage(Endpoints.ROLE_ICON(this.id, this.icon), undefined, undefined, this.guild.shard.client) : null;
     }
 
     /**
