+++ lib/structures/User.js
@@ -2,6 +2,7 @@
 
 const Base = require("./Base");
 const Endpoints = require("../rest/Endpoints");
+const formatImage = require("../util/formatImage");
 
 /**
 * Represents a user
@@ -98,7 +99,7 @@ class User extends Base {
         if(this.#missingClientError) {
             throw this.#missingClientError;
         }
-        return this.avatar ? this.#client._formatImage(Endpoints.USER_AVATAR(this.id, this.avatar)) : this.defaultAvatarURL;
+        return this.avatar ? formatImage(Endpoints.USER_AVATAR(this.id, this.avatar), undefined, undefined, this.#client) : this.defaultAvatarURL;
     }
 
     /**
@@ -112,7 +113,7 @@ class User extends Base {
         if(this.#missingClientError) {
             throw this.#missingClientError;
         }
-        return this.#client._formatImage(Endpoints.BANNER(this.id, this.banner));
+        return formatImage(Endpoints.BANNER(this.id, this.banner), undefined, undefined, this.#client);
     }
 
     /**
@@ -150,7 +151,7 @@ class User extends Base {
         if(this.#missingClientError) {
             throw this.#missingClientError;
         }
-        return this.avatar ? this.#client._formatImage(Endpoints.USER_AVATAR(this.id, this.avatar), "jpg") : this.defaultAvatarURL;
+        return this.avatar ? formatImage(Endpoints.USER_AVATAR(this.id, this.avatar), "jpg", undefined, this.#client) : this.defaultAvatarURL;
     }
 
     /**
@@ -166,7 +167,7 @@ class User extends Base {
         if(this.#missingClientError) {
             throw this.#missingClientError;
         }
-        return this.#client._formatImage(Endpoints.USER_AVATAR(this.id, this.avatar), format, size);
+        return formatImage(Endpoints.USER_AVATAR(this.id, this.avatar), format, size, this.#client);
     }
 
     /**
@@ -182,7 +183,7 @@ class User extends Base {
         if(this.#missingClientError) {
             throw this.#missingClientError;
         }
-        return this.#client._formatImage(Endpoints.BANNER(this.id, this.banner), format, size);
+        return formatImage(Endpoints.BANNER(this.id, this.banner), format, size, this.#client);
     }
 
     /**
